chore: updating folder structure
