+++ src/Properties/AssemblyInfo.cs
@@ -1,0 +1,10 @@
+﻿using System.Reflection;
+
+[assembly: AssemblyTitle("epi-display-samsung-mdc")]
+[assembly: AssemblyCompany("PepperDash Technology Corp")]
+[assembly: AssemblyProduct("PepperDashEssentials")]
+[assembly: AssemblyCopyright("Copyright © PepperDash Technology Corp 2020")]
+[assembly: AssemblyVersion("0.0.0.*")]
+[assembly: AssemblyInformationalVersion("0.0.0-buildtype-buildnumber")]
+[assembly: Crestron.SimplSharp.Reflection.AssemblyInformationalVersion("0.0.0-buildtype-buildnumber")]
+
