+++ src/SamsungMdcCommands.cs
