+++ src/SamsungMdcConfigObject.cs
