+++ src/SamsungMdcControllerFactory.cs
