+++ src/epi-display-samsung-mdc.nuspec
